+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/SimpleAuthorizationService.kt
@@ -24,9 +24,13 @@ import at.asitplus.openid.TokenIntrospectionRequest
 import at.asitplus.openid.TokenIntrospectionResponse
 import at.asitplus.openid.TokenRequestParameters
 import at.asitplus.openid.TokenResponseParameters
+import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import at.asitplus.signum.indispensable.josef.JsonWebKey
+import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.signum.indispensable.josef.JwsAlgorithm
+import at.asitplus.signum.indispensable.josef.JwsSigned
+import at.asitplus.wallet.lib.jws.VerifyJwsSignature
 import at.asitplus.wallet.lib.oidvci.CodeService
 import at.asitplus.wallet.lib.oidvci.CredentialIssuer
 import at.asitplus.wallet.lib.oidvci.DefaultCodeService
@@ -138,7 +142,9 @@ class SimpleAuthorizationService(
      */
     private val requestObjectSigningAlgorithms: Set<JwsAlgorithm.Signature>? = setOf(JwsAlgorithm.Signature.ES256),
     /** Used for [OAuth2AuthorizationServerMetadata.clientAttestationSigningAlgValuesSupportedStrings] */
-    private val supportedSigningAlgorithms: Set<JwsAlgorithm.Signature> = setOf(JwsAlgorithm.Signature.ES256)
+    private val supportedSigningAlgorithms: Set<JwsAlgorithm.Signature> = setOf(JwsAlgorithm.Signature.ES256),
+
+    private val trustedAttestationKeys: Set<CryptoPublicKey>? = null
 ) : OAuth2AuthorizationServerAdapter, AuthorizationService {
 
     private val _metadata: OAuth2AuthorizationServerMetadata by lazy {
@@ -418,6 +424,21 @@ class SimpleAuthorizationService(
         }
         val validatedClientKey = tokenService.verification.extractValidatedClientKey(httpRequest).getOrThrow()
 
+        httpRequest?.clientAttestation?.let {
+            val instanceAttestation = JwsSigned.deserialize<JsonWebToken>(
+                it = it,
+                deserializationStrategy = JsonWebToken.Companion.serializer(),
+            ).getOrElse {
+                throw InvalidGrant("unable to decode instance attestation: $it")
+            }
+            val instanceAttestationValid = trustedAttestationKeys?.firstNotNullOf { publicKey ->
+                VerifyJwsSignature().invoke(instanceAttestation, publicKey).isSuccess
+            } ?: false
+
+            if(!instanceAttestationValid) {
+                throw InvalidGrant("instance attestation invalid: $it")
+            }
+        }
         val clientAuthRequest = request.loadClientAuthnRequest(httpRequest, validatedClientKey)
             ?: throw InvalidGrant("could not load user info for $request")