+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidvci/ProofValidator.kt
@@ -1,13 +1,6 @@
 package at.asitplus.wallet.lib.oidvci
 
-import at.asitplus.openid.ClientNonceResponse
-import at.asitplus.openid.CredentialRequestParameters
-import at.asitplus.openid.CredentialRequestProofContainer
-import at.asitplus.openid.CredentialRequestProofSupported
-import at.asitplus.openid.IssuerMetadata
-import at.asitplus.openid.KeyAttestationRequired
-import at.asitplus.openid.OpenIdConstants
-import at.asitplus.openid.SupportedCredentialFormat
+import at.asitplus.openid.*
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.signum.indispensable.josef.JwsAlgorithm
@@ -15,10 +8,20 @@ import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.signum.indispensable.josef.KeyAttestationJwt
 import at.asitplus.wallet.lib.DefaultNonceService
 import at.asitplus.wallet.lib.NonceService
+import at.asitplus.wallet.lib.DefaultZlibService
+import at.asitplus.wallet.lib.agent.validation.StatusListTokenResolver
+import at.asitplus.wallet.lib.data.CredentialToJsonConverter.toJsonElement
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.RevocationListInfo
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusList
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListInfo
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
+import at.asitplus.wallet.lib.extensions.toView
 import at.asitplus.wallet.lib.jws.VerifyJwsObject
 import at.asitplus.wallet.lib.jws.VerifyJwsObjectFun
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.InvalidNonce
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.InvalidProof
+import kotlinx.serialization.json.Json
+import kotlinx.serialization.json.decodeFromJsonElement
 import kotlin.time.Clock
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.minutes
@@ -47,6 +50,8 @@ class ProofValidator(
     private val requireKeyAttestation: Boolean = false,
     /** Used to provide challenges to clients to include in proof of possession of key material. */
     private val clientNonceService: NonceService = DefaultNonceService(),
+
+    private val statusListTokenResolver: StatusListTokenResolver
 ) {
 
     /** Valid proof types for [SupportedCredentialFormat.supportedProofTypes]. */
@@ -112,7 +117,7 @@ class ProofValidator(
         val headerPublicKey = header.publicKey
             ?: throw InvalidProof("could not extract public key from $header")
 
-        return additionalKeys + headerPublicKey
+        return (additionalKeys + headerPublicKey).distinct()
     }
 
     /**
@@ -123,9 +128,23 @@ class ProofValidator(
         if (header.type != OpenIdConstants.KEY_ATTESTATION_JWT_TYPE) {
             throw InvalidProof("invalid typ: ${header.type}")
         }
-        if (payload.nonce == null || !clientNonceService.verifyNonce(payload.nonce!!)) {
-            throw InvalidNonce("invalid nonce: ${payload.nonce}")
+
+        val attestationStatus =
+            runCatching { Json.decodeFromJsonElement<StatusListInfo>(payload.status!!["status_list"]!!) }.getOrNull()
+                ?.let { statusListInfo ->
+                    val uri = statusListInfo.uri
+                    val statusListTokenPayload =
+                        runCatching { statusListTokenResolver(statusListUrl = uri) }.getOrNull()?.let {
+                            it.validate(revocationListInfo = statusListInfo, isInstantInThePast = { it < clock.now() })
+                        }
+                    val statusList = statusListTokenPayload?.getOrNull()?.revocationList as StatusList
+                    statusList.toView(DefaultZlibService()).getOrNull(statusListInfo.index)
+                }
+
+        if (attestationStatus != TokenStatus.Valid) {
+            throw InvalidProof("attestation revoked: $attestationStatus")
         }
+
         if (payload.issuedAt > (clock.now() + timeLeeway)) {
             throw InvalidProof("issuedAt in future: ${payload.issuedAt}")
         }
@@ -135,6 +154,7 @@ class ProofValidator(
         if (!verifyAttestationProof.invoke(this)) {
             throw InvalidProof("key attestation not verified: $this")
         }
+
         return payload.attestedKeys.map { it.toCryptoPublicKey().getOrThrow() }
     }
 }